requirements.txt
APScheduler==2.1.2
BeautifulSoup==3.2.1
Flask==0.10.1
Flask-Login==0.2.10
Flask-Mail==0.9.0
Flask-Migrate==1.2.0
Flask-Principal==0.4.0
Flask-RESTful==0.2.5
Flask-SQLAlchemy==1.0
Flask-Script==0.6.3
Flask-Security==1.7.4
Flask-WTF==0.9.5
Jinja2==2.7.2
MarkupSafe==0.18
Paste==1.7.5.1
SQLAlchemy==0.9.2
Werkzeug==0.9.4
aniso8601==0.82
arrow==0.4.2
boto==2.34.0
bottle==0.12.3
dnspython==1.11.1
ipaddr==2.1.11
ipython==1.2.0
itsdangerous==0.23
mock==1.0.1
nose==1.3.0
pika==0.9.13
psycopg2==2.5.2
python-dateutil==2.2
python-memcached==1.53
PyYAML==3.11
requests==2.2.1
py-bcrypt==0.4
Sphinx==1.2.2
gunicorn==18.0
M2Crypto==0.22.3
botocore==0.78.0
jira==0.32
